Replaces sizing screenshot
When creating a data api, the screenshot was taken from the beta.  This PR replaces that screenshot with the GA version that only allows for selection of 256mb size.  This is a temp fix to reduce potential exposure from customers who have bought from a marketplace as that purchase route is not billable right now with the new billing system.